🚚 rename: 스토리북 제작을 위해 폴더 구조 변경
