+++ apps/client/src/widgets/home/GuidesBox/GuidesBox.tsx
