+++ apps/client/src/widgets/home/WorkspaceContainer/WorkspaceContainer.tsx
