+++ apps/client/src/widgets/home/WorkspaceGrid/WorkspaceGrid.tsx
