+++ apps/client/src/widgets/home/WorkspaceHeader/WorkspaceHeader.tsx
