+++ apps/client/src/widgets/home/WorkspaceList/WorkspaceList.tsx
