+++ apps/client/src/widgets/home/WorkspaceModal/WorkspaceModal.tsx
